ze layout and add search functionality
Changes have been done in layout/Footer.tsx, layout/CardList.tsx, utils/config.ts, and pages/applications/index.tsx to improve the website's UI and add a search functionality. This was done to increase the application's usability and to enable users to easily locate specific content. Most notably, a color scheme was updated, logo responsiveness was enhanced, contact email was made as a clickable link for user convenience, and essential search functionality was added in CardList.