Standardize quotation marks, remove react-apollo
Applied changes to ensure consistency in the use of single quotes across different files. This makes the code more readable and adheres to the established coding standard. Removed the 'react-apollo' dependency from package.json as it is not being used anymore, which helps to reduce the size of the node_modules folder.