Update app UI and refactor codebase.
In this commit, various updates and refactoring have been done to the codebase. This includes theme updates, clean-up of unused code, and changes in code logic to allow it to be more maintainable. The UI of the application like the tables, forms, and banners were also improved to make the application look more professional. The 404 error page was also updated to include a 'Chrome Dino Game' for the user's entertainment. Lastly, updates were done to the routing setup and Apollo configurations for efficient data handling. These changes enhance the user experience and improve the code's readability and maintainability.