Implement play/pause feature for Slider and refactor element tags
Changes have been made to various components including layout and state behavior ones. Change has been made to the html element holding network chain name from 'p' tag to 'span' in the NetworkStatus.tsx. This makes it more semantically correct as tags should represent the type of content they hold and 'span' is more appropriate for short texts.

A new feature has been added that allows the user to play or pause the slider animation on Tokens slider. A new hook has been implemented that keeps the state of this feature, this hook is used across the Application.

The switch to control the animation has been added next to the Tokens title. On click it will toggle the state of the animation.

These changes improve user interaction in the application by providing users with more control over the UI.