Replace logo with dynamic AppLogo component
This commit introduced a new AppLogo component that dynamically selects a logo source based on the current domain. It also modified the Header, Footer and Loading components to use the new AppLogo component instead of the previous hard-coded logo. This is useful for instances where the application is deployed under different domains and each requires a unique logo. Added getFirstLevelDomain utility function to support this feature.