Update application to handle dynamic favicons and online status display
Added functionality to handle dynamic favicons based on the hostname in `Head.tsx`. Icons for localhost, byte.exchange, and metamask.digital were included. Introduced a new component `IconHandling` to link the chosen favicon to the head of the document. This change enhances the website's appearance at the tab of the browser and ensures consistency across different hostnames.

Created a custom react hook in `OnlineStatus.tsx` to check for the online status in real-time of the application. This was then used in the `Header.tsx` to display an 'offline' message to the user when the network is made unavailable.

Also made minor bug fixes such as adding missing type to the icon object in `AppLogo.tsx` and adjusting the draggable attribute in `Loading/index.tsx`. These additions increase the reliability and user-friendliness of the application.

Commented out the static favicon link in `_document.tsx` to allow for dynamic assignment of the icons in `Head.tsx`. This makes the application more scalable and adaptable to different hostnames.