Refine UI elements for enhanced user experience
This commit includes several tweaks to the user interface to provide a better user experience.
In the Header, the positioning of page navigation is made sticky for easier navigation, and a blur effect is added to the background for enriched visual effect. The size of the AppLogo is reduced for a cleaner and less intrusive appearance.
In the Footer, the AppLogo is removed to decrease clutter, and the font is changed to enhance readability.
In Transaction, the size of search text input and dropdown selectors are reduced to ensure a tidy and coordinated look.
In Token, a 'View All' button is added for user navigation and to increase discoverability of additional content. The Card backdrop is given a blur effect to match the aesthetic in other components. All these changes were pursued with the objective to enhance usability and aesthetic appeal overall.